continious damage turret, turret hierarchy refactoring

+++ TowerDefence/Assets/TowerDefence/Scripts/EnemyLogic/HealthLogic/HealthComponent.cs
@@ -6,9 +6,9 @@ namespace TowerDefence.Scripts.EnemyLogic.HealthLogic
 	public abstract class HealthComponent : MonoBehaviour
 	{
 		[SerializeField]
-		protected int _health;
+		protected float _health;
 
 		public Action OnZeroHealth;
-		public abstract void ReduceHealth(int value);
+		public abstract void ReduceHealth(float value);
 	}
 }
